requirements.txt
streamlit==1.32.2
langchain-google-genai==0.0.6
faiss-cpu==1.7.4
python-dotenv==1.0.1
pypdf==4.1.0
sentence-transformers==2.6.1
google-generativeai==0.3.2
numpy
pytesseract
Pillow
PyMuPDF